Use PSK compatiable cipher suite on server
Use configured cipher, or default to a SHA-256 one.

Note: here we've hard-coded PTLS_CIPHER_SUITE_AES_128_GCM_SHA256 for the default,
but any SHA256 one could be chosen (unless the client is sending early data,
but in that case both endpoints need to specify it explicitly, not rely on default).